warning fix

+++ ugbase/pcl/parallel_archive.cpp
@@ -28,7 +28,7 @@ void WriteParallelArchive(ProcessCommunicator &pc, std::string strFilename, cons
 	MPI_Comm m_mpiComm = pc.get_mpi_communicator();
 	MPI_File fh;
 
-	bool bLast = pc.get_local_proc_id()+1 == pc.size();
+	bool bLast = pc.get_local_proc_id()+1 == (int)pc.size();
 	bool bFirst = pc.get_proc_id(0) == pcl::GetProcRank();
